move code out of preindex/__init__.py
